requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2
djangorestframework==3.14.0
dnspython==2.3.0
jsonschema==4.17.3
mongoengine==0.27.0
pymongo==4.3.3
pyrsistent==0.19.3
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
